# Conflicts:
#	app/package.json
#	app/src/terminal/Terminal.svelte
#	app/src/tutorials/Tutorial.svelte

+++ app/build.sh
@@ -4,7 +4,7 @@
 echo "<!-- Build: $(TZ=America/Los_Angeles date); $(git rev-parse HEAD) -->" >> public/index.html
 
 # Create mock routes (only used for local testing; there's a router server-side to handle this)
-PAGES=("tutorials" "playground" "rosalind")
+PAGES=("tutorials" "playground" "sandboxes" "rosalind")
 for page in ${PAGES[@]}; do
 	mkdir -p "public/$page"
 	cp public/index.html public/$page/index.html
@@ -14,6 +14,7 @@ done
 mkdir -p public/data
 for tutorial in $(ls -d src/tutorials/*/);
 do
+	[[ ! -e $tutorial/data ]] && continue;
 	dest=public/data/$(basename $tutorial)
 	mkdir -p $dest
 	cp $tutorial/data/* $dest
